JPF: Make ftp commands get executed in a different thread to the one which runs the control connection, use a threadpool to avoid thread creation overheads.
A consequence of this is that it is possible to correctly implement ABOR support.

This has been tested and no problems were noticed but it could certainly benefit from more widespread testing, so please play and report any problems.

There is currently debug logging of thread pool details after each command is executed, this will be removed at some point but is there now to keep an eye out for thread leakage in the current code. Please report any unusual values you see in this, (pool size should not exceed 2 in normal operation)

git-svn-id: http://svn.code.sf.net/p/drftpd/svn/branches/jpf@1692 c83153ee-f3e8-0310-b165-dc3a72e09794