Add a partial remerge option in an attempt to speed up slave remerges, this still won't by any means be instantaneous as the slave still has to traverse its entire filesystem, it just means that messages to the master (and the corresponding inode comparison on the master) can be skipped for much of the filesystem. This will provide two options for partial remerges in addition to retaining the full remerge capability along with a per slave mechanism to override the master requesting partial remerges. See the changes to master.conf and slave.conf for the new settings, the default in the repository for now will be to have these options disabled.
This mechanism is unable to cope with many scenarios of moving directories around locally on a slave so after performing such operations a full remerge should always be performed to synchronise the VFS (the slave.conf override is primarily intended for this purpose), this mechanism is simply to try and achieve faster reconnects during normal operation of a site.
Additionally this mechanism has no impact on a forced "site remerge" operation which will continue to always perform a full merge of the requested directory tree.

Although this has been tested and appears to work fine, as usual proceed at your own risk!

git-svn-id: http://svn.code.sf.net/p/drftpd/svn/trunk@2041 c83153ee-f3e8-0310-b165-dc3a72e09794