Introduce a mutex between getting a checksum from a slave for a file and deleting the file, this prevents the file being deleted in another thread whilst the checksum request is in progress, previously this would kick the slave offline when trying to update the new checksum in the VFS on return.
Alter the LIST/MLST commands so that if an entry is deleted in the middle of building the list output this entry will simply be omitted from the list rather than returning a potentially malformed response as the previous code would. This adds a slight overhead to the construction of the output (using a StringBuilder per line) but should not be significant.

git-svn-id: http://svn.code.sf.net/p/drftpd/svn/trunk@2108 c83153ee-f3e8-0310-b165-dc3a72e09794