Resolve an issue with ghost FTP connections remaining and errors about removing connections appearing in wrapper-master.log, this fixes #294.
For the interested the following is an explanation of the cause of this bug:
The BaseFTPConnection objects are stored in a Vector, these are added prior to being executed in the ThreadPool and are removed when the BaseFTPConnection.run() method ends or if execution is rejected by the ThreadPool.
The implementation of Vector.remove() in the JDK is as follows:
 * Find the index of the requested object in the elements array.
 * If an index was found then remove the object currently in this index from the array and return true otherwise if no index was found return false.

To find the index in the elements array the implementation iterates the array testing the equality of each element, essentially elements[i].equals(objectToFind) and returns the current index if this expression is true.

So the important bit is what actually happens when calling BaseFtpConnection.equals(anotherObject), to see what happens we follow the class hierarchy through the superclasses until one implementing an equals() method is found, the class hierarchy of BaseFTPConnection is as follows:
 * BaseFTPConnection
 * Session
 * KeyedMap<K,V>
 * Hashtable<K,V>

This is as far as it needs to go to find an equals() implementation as one is provided by the JDK Hashtable class, this class returns true if the references of the objects are the same or if both Maps are the same size and contain exactly the same set of key/value pairs, importantly this is true for empty maps also, two empty Hashtable instances will return true for mapA.equals(mapB).

Now until some entries are populated in the session keyedmap for the BaseFTPConnection object the connection is essentially an empty Hashtable, this is particularly true on new connections and those rejected due to bans, login restrictions etc (and also on ones that are rejected by the ThreadPool and thus removed before any user commands are processed).

If for example there are currently two such connections at indexes 3 and 5 in the Vector of BaseFTPConnection objects and the connection at index 5 terminates first when removing this connection we actually remove the connection at index 3 since the Vector removes the *first* object which is equal in its array. At a later stage we want to remove the connection which was originally at index 3 when it terminates, it may seem intuitive that this will match an equality test with the ghost connection in index 5 and prevent an issue, however between the time the connection in index 3 was wrongly removed and actually wanting to remove the connection in index 3 the keyedmap may have been populated, at this point the equality test fails. We are now left with an error trying to remove the connection originally in index 3 and a permanent ghost connection in index 5. QED :-)

The fix for this is to provide an overriden BaseFTPConnection.equals() which simply checks equality of references and therefore will only return true when the objects being compared are in fact the same object.

git-svn-id: http://svn.code.sf.net/p/drftpd/svn/trunk@2323 c83153ee-f3e8-0310-b165-dc3a72e09794