New TrialManager Plugin.  This will allow different trial types to be loaded/unloaded cleanly without duplicate code from 1 configuration file.  This is similar to how Archive works.
This is just the manager right now, as most (if not all) of the types will not be held in trunk but on the forums due to maintance factor.

Currently there are only 3 commands (TOP/CUT/PASSED) which can be used by any of the types.  Other commands can be added through a seperate manager for different types.

git-svn-id: http://svn.code.sf.net/p/drftpd/svn/trunk@2400 c83153ee-f3e8-0310-b165-dc3a72e09794