Added the rest of the minimum speed limit stuff.  The architecture is now in place for the plugins to work with this.
Using the keys Dom added in [2414] and changing the slave messages to an array plus modifications to the slave transfer loop in this commit will allow for min/max speed to be set on a prehook of RETR/STOR before getting sent to the slave.  If the speed falls below X, it will throw a SlowTransferEvent which can be handled by the plugin.

This finishes the core modifications of #131.

git-svn-id: http://svn.code.sf.net/p/drftpd/svn/trunk@2434 c83153ee-f3e8-0310-b165-dc3a72e09794