Changed privpath behavior to check against all inodes and not just directories. This according to #169
Removed regex check for privpath on lucene search, if too many privpath rules are added lucene search crawls on its knees.
It now returns all results found in index and the check for hidden files are handled when looping the results.
I left the method to create a regex for all privpath rules in case its needed elsewhere or later.

Some minor bugfixes was addressed as well.


git-svn-id: http://svn.code.sf.net/p/drftpd/svn/trunk@2499 c83153ee-f3e8-0310-b165-dc3a72e09794